Do not kill gpgagent just yet when logged in other shells

+++ .bash_logout
@@ -1,2 +1,4 @@
-killall -u $(whoami) gpg-agent &>/dev/null
+if [[ $(w -h | wc -l) -le 2 ]] ; then
+    killall -u $(whoami) gpg-agent &>/dev/null
+fi
 history -a &>/dev/null
